Move ServerCookieEncoder and use it
