+++ components/api/src/main/java/com/hotels/styx/api/cookies/ResponseCookie.java
@@ -22,7 +22,6 @@ import com.hotels.styx.api.HttpHeaders;
 import io.netty.handler.codec.http.cookie.ClientCookieDecoder;
 import io.netty.handler.codec.http.cookie.Cookie;
 import io.netty.handler.codec.http.cookie.DefaultCookie;
-import io.netty.handler.codec.http.cookie.ServerCookieEncoder;
 
 import java.util.ArrayList;
 import java.util.Collection;
