+++ components/api/src/test/java/com/hotels/styx/api/cookies/ServerCookieEncoderTest.java
@@ -13,7 +13,7 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-package com.hotels.styx.server.netty.codec;
+package com.hotels.styx.api.cookies;
 
 import io.netty.handler.codec.http.cookie.ClientCookieDecoder;
 import io.netty.handler.codec.http.cookie.Cookie;
